Avoid an assert inside avahi
The conditions leading to the assert are documented in the comments. I
found this while porting sane-airscan to openbsd, but I don't see
anything openbsd specific about the issue. Maybe avahi on linux is
built with -DNDEBUG?

This is the simple fix, but it is, theoretically at least, inefficient
since now we block for a millisecond at a time. In practice, a
scanimage is as fast as it was before on linux.

Probably the ideal fix would be in avahi. It could just ignore a
watcher if w->idx < 0. It would be handled in the next call to
avahi_simple_poll_iterate.

I am sending this patch in order to get airscan working with a current
avahi version.