Comment out code depending on IP_PKTINFO
This is a hack, but the last patch to get airscan working on OpenBSD
on my laptop. Maybe only because it has only one interface.

Looking at the openbsd source code, I don't see a way to get the
interface where a ipv4 packet was received.

I am not familiar with this code, but it looks like each resolver has
a different socket. Could we map directly from file descriptor to
resolver?

Signed-off-by: Rafael Ávila de Espíndola <rafael@espindo.la>