Allow the avahi loop to sleep longer
This is optional, and I am not sure the complexity is worth it.

With this patch the avahi loop can block in a poll for as much as it
wants. The flip side is that we have to wake it up when we want to
grab eloop_mutex.

This is done by having a thread that continuously wakes it up. To
limit the busy work, that thread is only active when requested by the
main thread and the avahi loop is actually in a poll.

I was able to do this without an extra thread for eloop_mutex_lock, by
using pthread_mutex_trylock, but could not figure out how to do it for
eloop_cond_wait.

A disadvantage this has is that in

    pthread_mutex_unlock(&wakeup_mutex2);
    rc = poll(ufds, nfds, timeout);
    pthread_mutex_lock(&wakeup_mutex2);

The lock might block until eloop_wakeup_thread_func grabs wakeup_mutex
and goes for another round.

I tried fixing the above by using a condition variable, but as far as
I can tell I would have to make eloop_mutex non recursive, which is a
bigger change than I would like for this.